move external data loading to own file
